predecessor.py in own module; add resolver to drill into predecessor objects;
